WI #2225 Delete sql_column_name file
